// src/types/slices/rrhhType.ts
export type RRHH = {
    id: number
    n_document: string
    name: string
    surname: string
    second_surname: string
    birth_date: string
    sexo: number
    direccion: string
    iddistrict: number
    idcharge: number
    charge: string
    idauditoria: number
    idcompany: number
    status: string
}